ticks: move breakpoints code into a separate location
